Implement authentication via itch.
itch.io is an online marketplace and app for playing games. It is a
popular choice for FOSS games, so I think there could be a fair bit of
crossover with Nakama.

Games started via the itch app are passed a JWT token that can be used
to authenticate the user with the itch API. The idea is that the game
passes this token to the server, and the server passes the token to
itch.

This feature was discussed on the forums at
https://forum.heroiclabs.com/t/itch-io-authentication/

Depends on: https://github.com/heroiclabs/nakama-common/pull/9/files

See also:
- https://itch.io/docs/itch/integrating/manifest.html
- https://itch.io/docs/api/serverside
- https://github.com/itchio/itch.io/issues/536#issuecomment-274472130