fix list matches behavior with no querystring
An issue was reported in which matching on the label string
was not returning expected matches.  This was diagnosed to relate
to the use of a MatchQuery clause.  When using TermQuery the
correct results were returned.

Further analysis revealed that the existing unit tests were not
covering this path.  The
TestMatchRegistryAuthoritativeMatchAndListMatches test was
intended to cover this path, but did not due to passing an
empty string instead of a nil query string.  This commit updates
this test to correctly exercise this code path.

Further, a new test case has been added.  This test matches
TestMatchRegistryAuthoritativeMatchAndListMatches but also uses
a label string which is able to be tokenized (if a Match query
is used and a tokenizer is incorrectly applied).  This test
case reproduces the original issue, where a label that can
be tokenized results in the MatchQuery returning incorrect
results.

The original bug has been addressed by switching to using a
TermQuery.