enable boosting with matchmaker
While enabling boost for the Matchmaker a slight issue was
observed.  Our goal is to support the use of boost to influence
scores in a particular way, but implicit in that was that if
a query does not use boost, we expect the previous behavior
(in which items were sorted by create-time).

This is not quite as simple as it sounds, as previously the
scores were not being used at all, so even if some hits had
different scores, it did not matter.  This commit adds a unit
test for the matchmaker which highlights this issue.  It does
not use any boost, but the scores vary slightly, resulting in
a change in matchmaking behavior.

To help address this issue, we have added additional Bluge
configuration to specify a DefaultSimilarity to be used
instead of BM25.  This will have the effect of scoring all
text matches as 1.0 * boost, instead of a meaningless BM25
calculation.