initial support for mutual matchmaking
This commit introduces 3 unit tests, each of which allow for
a different type of undesired non-mutual match.  It also fixes
one existing unit test which was previously passing in spite of
an undesirable non-mutual match.

The matchmaking changes attempt to make minimal changes to the
existing logic, and simply try to catch mutual matchmaking
violations, and disallow them during the match building
process.