update match tests to not rely on time.Sleep
The purpose of this commit is to improve the match related tests
to manually drive the update/processing logic, and not rely on
time.Sleep to allow for "some" processing to occur.

To do this, the MatchRegistry and Matchmaker are updated to
allow for a config value of 0 for the label update interval
and for the matchmaking interval.  When configured with such
a value, they will not automatically process data, instead
they rely on manual invocation of their process methods.

Finally, the tests are updated to configure these components with
the 0 value, and then in test logic, sleep statements are replaced
with the appropriate call to invoke the processing logic once.