fix Matchmaker enforcement of max tickets
The Matchmaker tracks tickets by session and party for the
purposes of enforcing a MaxTickets configuration setting.  It
has been observed that in some cases, we attempt to remove
incorrect entries, which results in the correct entries leaking
inside the tracking maps.  In addition to leaking resources,
this could cause incorrect behavior for the affected session
and party identifiers.

Two unit tests have been added to illustrate the problem and
validate the fix.  These tests fail without the fix on a
random basis, and it is recommended to run them with
-count=100 to ensure the problematic scenario is encountered.