Remove lock from sessionWS.SendBytes
There are six different calls to s.Lock() in sessionWS.  Three of these locks are around calls to SetWriteDeadline and WriteMessage on the WebSocket.  WriteMessage will block for up to the timeout set in SetWriteDeadline, which has a default of 5000 milliseconds.  This lock is also acquired by SendBytes to determine if the session has been stopped.  This is a problem as the event processing methods in StatusRegistry and Track call SendBytes directly on the Session.  If one of the Sessions being written to has delays in writing data, this could cause backups in the event processing.

The simplest approach (this commit) is to just remove the lock within SendBytes.  To avoid a race condition with writing to a closed channel, we need to not close the outgoingCh channel in the Close method.  This is safe to do as channels do not need to be explicitly closed and will automatically be cleaned up by the garbage collector.  It should also be possible to remove the lock entirely from sessionWS, but that is a much larger change.