NK-589 Add Party Search support (#2333)
Ensure that an error is returned if Google Auth is used without proper configuration.
Improve Google auth error handling and message.
Do not redact empty config values to be able to differentiate if values are set or not in the console configuration page.
Allow to set a party label on creation.
Add PartyUpdate socket function to update the label.
Add PartyList REST and runtime API.